changed the way the potentialResultSize is computed, this avoids skipping some enumerator elements because the featureList suggests a possibly bug number of variants whereas the number of individual features used is quite low
changed environment used to add variables in CDeclUse in addFunction, now using the updated Env instead of the old Env:
this fixes a particular bug where you declare a function with variable modifiers, redeclare it in one specific context. Before this change we used to remove the original definition in the declUseMap for the new redeclaration, although it is only a redeclaration under one specific context.